Add Request() method to structs
In addition, moved the structs to seperate files to closely match the
layout as defined in the RFC23 in the table of contents where it lists
the various tags.
