+++ hls/types-segment.go
@@ -1,0 +1,136 @@
+package hls
+
+import (
+	"fmt"
+	"net/http"
+	"time"
+)
+
+//Segment represents the Media Segment and its tags
+type Segment struct {
+	ID              int //Sequence number
+	URI             string
+	Inf             *Inf //Required.
+	Byterange       *Byterange
+	Discontinuity   bool //Represents tag #EXT-X-DISCONTINUITY. MUST be present if there's change in file format; number, type and identifiers of tracks or timestamp sequence
+	Keys            []*Key
+	Map             *Map
+	ProgramDateTime time.Time //Represents tag #EXT-X-PROGRAM-DATE-TIME
+	DateRange       *DateRange
+
+	mediaPlaylist *MediaPlaylist // MediaPlaylist is included to be used internally for resolving relative resource locations
+}
+
+// Request creates a new http request ready to send to retrieve the segment
+func (s *Segment) Request() (*http.Request, error) {
+	uri, err := resolveURLReference(s.mediaPlaylist.URI, s.URI)
+	if err != nil {
+		return nil, fmt.Errorf("failed building resource url: %v", err)
+	}
+
+	req, err := http.NewRequest("GET", uri, nil)
+	if err != nil {
+		return req, fmt.Errorf("failed to construct request: %v", err)
+	}
+	return req, nil
+}
+
+// Segments implements golang/sort interface to sort a Segment slice by Segment ID
+type Segments []*Segment
+
+func (s Segments) Len() int {
+	return len(s)
+}
+func (s Segments) Swap(i, j int) {
+	s[i], s[j] = s[j], s[i]
+}
+func (s Segments) Less(i, j int) bool {
+	return s[i].ID < s[j].ID
+}
+
+// Inf represents tag
+// 		#EXTINF: <duration>,[<title>]
+type Inf struct {
+	Duration float64
+	Title    string
+}
+
+// Byterange represents tag #EXT-X-BYTERANGE.
+// Introduced in HLSv4.
+// Format: length[@offset].
+type Byterange struct {
+	Length int64
+	Offset *int64
+}
+
+// Key represents tags #EXT-X-KEY:<attribute=value> and #EXT-X-SESSION-KEY. Specifies how to decrypt an encrypted media segment.
+// #EXT-X-SESSION-KEY is exclusively a Master Playlist tag (HLS V7) and it SHOULD be used if multiple Variant Streams use the same encryption keys.
+// TODO(jstackhouse): Split SESSION-KEY into it's own type as it's got different validation properties, and is part of the master playlist, not media playlist.
+type Key struct {
+	IsSession         bool   //Identifies if #EXT-X-KEY or #EXT-X-SESSION-KEY. If #EXT-X-SESSION-KEY, Method MUST NOT be NONE.
+	Method            string //Required. Possible Values: NONE, AES-128, SAMPLE-AES. If NONE, other attributes MUST NOT be present.
+	URI               string //Required unless the method is NONE. Specifies how to get the key for the encryption method.
+	IV                string //Optional. Hexadecimal that specifies a 128-bit int Initialization Vector to be used with the key.
+	Keyformat         string //Optional. Specifies how the key is represented in the resource. V5 or higher
+	Keyformatversions string //Optional. Indicates which Keyformat versions this instance complies with. Default value is 1. V5 or higher
+
+	mediaPlaylist *MediaPlaylist // MediaPlaylist is included to be used internally for resolving relative resource locations
+}
+
+// Request creates a new http request ready to retrieve the segment
+func (k *Key) Request() (*http.Request, error) {
+	uri, err := resolveURLReference(k.mediaPlaylist.URI, k.URI)
+	if err != nil {
+		return nil, fmt.Errorf("failed building resource url: %v", err)
+	}
+
+	req, err := http.NewRequest("GET", uri, nil)
+	if err != nil {
+		return req, fmt.Errorf("failed to construct request: %v", err)
+	}
+	return req, nil
+}
+
+//Map represents tag #EXT-X-MAP:<attribute=value>. Specifies how to get the Media Initialization Section
+type Map struct {
+	URI       string     //Required.
+	Byterange *Byterange //Optional. Indicates the byte range into the URI resource containing the Media Initialization Section.
+
+	mediaPlaylist *MediaPlaylist // MediaPlaylist is included to be used internally for resolving relative resource locations
+}
+
+// Request creates a new http request ready to retrieve the segment
+func (m *Map) Request() (*http.Request, error) {
+	uri, err := resolveURLReference(m.mediaPlaylist.URI, m.URI)
+	if err != nil {
+		return nil, fmt.Errorf("failed building resource url: %v", err)
+	}
+
+	req, err := http.NewRequest("GET", uri, nil)
+	if err != nil {
+		return req, fmt.Errorf("failed to construct request: %v", err)
+	}
+	return req, nil
+}
+
+//DateRange represents tag #EXT-X-DATERANGE:<attribute=value>.
+//
+//If present, playlist MUST also contain at least one EXT-X-PROGRAM-DATE-TIME tag.
+//Tags with the same Class MUST NOT indicate ranges that overlap.
+type DateRange struct {
+	ID               string    //Required. If more than one tag with same ID exists, att values MUST be the same.
+	Class            string    //Optional. Specifies some set of attributes and their associated value semantics.
+	StartDate        time.Time //Required.
+	EndDate          time.Time //Optional.
+	Duration         *float64  //Optional. If both EndDate and Duration present, check EndDate equal to Duration + StartDate
+	PlannedDuration  *float64  //Optional. Expected duration.
+	XClientAttribute []string  //Optional. Namespace reserved for client-defined att. eg. X-COM-EXAMPLE="example".
+	EndOnNext        bool      //Optional. Possible Value: YES. Indicates the end of the current date range is equal to the start date of the following range of the samePROGRAM-DATE-TIME class.
+	SCTE35           *SCTE35
+}
+
+//SCTE35 represents a DateRange attribute SCTE35-OUT, SCTE35-IN or SCTE35-CMD
+type SCTE35 struct {
+	Type  string //Possible Values: IN, OUT, CMD
+	Value string //big-endian binary representation of the splice_info_section(), expressed as a hexadecimal-sequence.
+}
